move write functions into separate script
