--- etc/write_ipmi
@@ -1,16 +1,0 @@
-#!/usr/bin/env python
-
-# Writes data piped from stdin to EEPROM...good luck
-
-from sys import stdin
-
-from i2c import eeprom
-
-byte_data = stdin.buffer.read()
-
-bus = smbus2.SMBus(0)
-if eeprom.detect_16bit():
-    eeprom.write_16bit_address(byte_data)
-else:
-    eeprom.write_8bit_address(byte_data)
-
